Translate and update make_asym function
Translate the old c function into c++. This simplifies the wrapping
and compiling as pybind11 can be used, so the code can stay pure c++
and have pybind11 handle the interfacing, no python c api needed in
the algorithm itself. In the translation of the code, the algorithm
was modified so as to not need a set of double for loops. The
modification was possible by making use of some algebraic transforms.
This change decreased the run time by about 50%