Speed up asymetry value calculation
Modify the asymmetry value calculation to eliminate extra for loops.
This is done by calculating the radius bins on the fly from the
indexes of the data array, and doing the book keeping in a smaller
array the size of the number of unique radii. The python code was
updated to the change in interface of the make_asym function,
that is to remove the creation of the distance array. To maintain
readability for future developers, the pure python make_asym
function was updated to the same function call signature, but it
internally creates a distance array. This python function is not
optimized for speed, but instead optimized for the readability/
understandability from a python programmer point of view. Anyone
needing a fast runtime should use the c++ extension. After the
modifications the runtime for the routine was brought down to
about 5ms, down from 30ms during the switch to pybind11, and 50ms
in the pure c implementation.