Provide better feedback on error
When the make_asym function is fed arrays which are not square it
will throw a RuntimeError. In the normal course of running asym
this is normally because the centering routine is running off the
edge of the input image due to a problem in the image, or an
almost flat asymmetry space (uniform distribution in the image).
Catch the error thrown by make_asym, and re-throw a more appropriate
message for the context of the centering function.